frontend_Dockerfile_and_other/requirements.txt
requests==2.25.1
flask==2.0.2
prometheus_flask_exporter==0.18.7
mysql-connector-python==8.0.28
mysqlclient==1.4.4
python-dotenv==0.19.2
